+++ gradle.properties
@@ -9,8 +9,8 @@ minecraft_version               = 1.21
 # These should be automatically updated, unless the environment
 #  variable "DISABLE_PROPERTIES_UPDATE" is set.
 yarn_version                    = 1.21+build.9
-loader_version                  = 0.16.0
-fapi_version                    = 0.100.7+1.21
+loader_version                  = 0.16.7
+fapi_version                    = 0.102.0+1.21
 
 # Dependencies
 ## DevAuth, check at https://github.com/DJtheRedstoner/DevAuth
@@ -18,10 +18,10 @@ devauth_version                 = 1.2.1
 
 monkeylib538_version            = 2.0.1+1.21
 
-jgit_version                    = 6.9.0.202403050737-r
+# MESH Lib, check at https://github.com/OffsetMods538/MESH-Lib
+mesh_lib_version                = 1.0.0+1.21
 
-# Netty, currently set to the version of netty used in 1.19.2, hopefully won't cause problems
-netty_version                   = 4.1.77.Final
+jgit_version                    = 6.9.0.202403050737-r
 
 undertow_version                = 2.3.14.Final
 jakarta_servlet_api_version     = 6.0.0
