--- src/main/java/top/offsetmonkey538/githubresourcepackmanager/utils/HttpUtils.java
@@ -1,27 +1,0 @@
-package top.offsetmonkey538.githubresourcepackmanager.utils;
-
-import io.netty.buffer.ByteBuf;
-import io.netty.buffer.Unpooled;
-import io.netty.channel.ChannelFutureListener;
-import io.netty.channel.ChannelHandlerContext;
-import io.netty.handler.codec.http.DefaultFullHttpResponse;
-import io.netty.handler.codec.http.FullHttpResponse;
-import io.netty.handler.codec.http.HttpResponseStatus;
-import io.netty.util.CharsetUtil;
-
-import static io.netty.handler.codec.http.HttpHeaderNames.CONTENT_TYPE;
-import static io.netty.handler.codec.http.HttpVersion.HTTP_1_1;
-
-public final class HttpUtils {
-    private HttpUtils() {}
-
-
-
-    public static void sendError(ChannelHandlerContext ctx, HttpResponseStatus status) {
-        final ByteBuf byteBuf = Unpooled.copiedBuffer("Failure: " + status.toString() + "\r\n", CharsetUtil.UTF_8);
-        final FullHttpResponse response = new DefaultFullHttpResponse(HTTP_1_1, status, byteBuf);
-
-        response.headers().set(CONTENT_TYPE, "text/plain; charset=UTF-8");
-        ctx.writeAndFlush(response).addListener(ChannelFutureListener.CLOSE);
-    }
-}
