+++ src/main/resources/github-resourcepack-manager.mixins.json
@@ -5,7 +5,6 @@
 	"mixins": [
 		"AbstractPropertiesHandlerAccessor",
 		"MinecraftDedicatedServerAccessor",
-		"ServerNetworkIoMixin",
 		"ServerPropertiesLoaderAccessor"
 	],
 	"injectors": {
